Fetch stables on pull
Summary:
When running pull, also refetch stable locations. This will update the repo's stable locations, so the next hg log will include them. By doing this when we pull, it's a strong user signal that they want to have the latest value for the stable, so it's a good time to keep it up to date.

We do this regardless of which pull button choice you make (arc pull and sl pull)

Reviewed By: quark-zju

Differential Revision: D55453805

fbshipit-source-id: d682669fe9a920ea1789b813f02753befb0eae55