Attach fetched stable locations as CommitInfo.stableCommitMetadata
Summary:
After setting repo.stableLocations, we'll get the appropriate public commits in the results from `log`. However, the fields in CommitInfo won't know anything about the stable locations.

We could handle this by having the UI layer know the stable locations (which it will anyway), and choose to specially render them. However, we get a lot of stuff "for free" if we just set the CommitInfo.stableCommitMetadata. It's already named correctly for this use case and everything, so I say we just use this part of CommitInfo. It's maybe not completely obvious, since this data is not visible in CLI `sl log` or `sl sl`. But I think that's OK. Stable locations is an ISL thing right now.

Reviewed By: quark-zju

Differential Revision: D55342081

fbshipit-source-id: 29aea54714136cadd0a4f52d8ef236541754f163