Consolidate <Bookmark> to support stableCommitMetadata
Summary:
We have a <Bookmarks> component that renders a list bookmarks. It's basically the same as how we render stableCommitMetadata "bookmarks", so let's consolidate the implementation to allow you to pass either simple bookmark names or stableCommitMetadata entries. This allows us to show tooltips on stableCommitMetadata entries. We also render stableCommitMetadata values slightly differently as a hint that they're different from normal bookmarks.

Also, our dag render subset logic was capable of stripping out commits with stableCommitMetadata. We should treat it like a remoteBookmark and always show it if it has non-empty data there.

Reviewed By: quark-zju

Differential Revision: D55342079

fbshipit-source-id: 37f2246bb627dac86782b6cdafa6a1faef2b41df