Filter useless public commits from dag
Summary:
(updated from feedback)

Update the CommitInfo's fed to the `latestDag` to not include hidden bookmarks. This way, all consumers downstream using the dag will not include any bookmarks we don't care about.

This makes it automatic to not include hidden bookmarks in any logic using bookmarks.

In a follow up diff, we can actually clean up special handling of hidden bookmarks.

This diff also means public commits with no bookmarks after hiding will be hidden thanks to dag's render subset.

Note that we do need the bookmark manager to see all the original bookmarks, so it has to not use the dag when rendering and rather use the raw commit data. I guess we could have a separate dag selector which does this transformation but that feels unnecessary.

Reviewed By: quark-zju

Differential Revision: D55489711

fbshipit-source-id: da4e3582f2a14270b55148031c6003532548a508