Simplify suggestedRebaseDestinations
Summary:
Now that we filter remote bookmarks at the dag layer, we no longer need downstream users to filter hidden bookmarks. This means we can remove that filtering from our suggested rebase destinations.

However, the filtering is done on the dag, so we need to use latestDag instead of latestCommits. I think this is actually preferrable though.

Reviewed By: quark-zju

Differential Revision: D55495382

fbshipit-source-id: d4693f2f70ba9ecbaf8ce5894d4ae58a57696c84