[bug] remove decode() because type is str not bytes
FYI, I discovered this because I tried using the devcontainer
stuff introduced in #807, so from the Docker container launched
by VS Code, I tried running:

```
./tests/run-tests.py ./tests/test-doctest.py
```

It prints out the following:

```
WARNING: Did not find prerequisite tool: unzip
```

I guess that means `unzip` doesn't get installed as part
of the Docker container, though I suppose it isn't a big deal?
That is, this is just a warning and the test passes just fine.