FUSE: add logging at channel creation time
Summary:
# Context

The number of FUSE dispatcher threads to spawn on mount creation currently defaults to 16. To change this value, the user would have to pass in a custom argument to the Eden daemon. This is not easy to do since the EdenFS CLI handles daemon startup.

We should be able to configure the number of FUSE dispatcher threads without instructing users to manually spawn an Eden daemon. To make this easier, we can just configure the number of threads with an Eden daemon config option. We can then later edit this config in real time w/ Eden dynamic config.

Raising this value may help with workflows that do a lot of IO in parallel (ex: running many ripgreps at once). This might also be used as a potential mitigation for S402137.

# This diff

Makes it easier to see what fuse channel options are configured at mount creation time.

Reviewed By: genevievehelsel

Differential Revision: D55496101

fbshipit-source-id: b5914153ed22a812803c842b4f5378e70655bc50