FUSE: add FUSE dispatcher threads config
Summary:
# Context

The number of FUSE dispatcher threads to spawn on mount creation currently defaults to 16. To change this value, the user would have to pass in a custom argument to the Eden daemon. This is not easy to do since the EdenFS CLI handles daemon startup.

We should be able to configure the number of FUSE dispatcher threads without instructing users to manually spawn an Eden daemon. To make this easier, we can just configure the number of threads with an Eden daemon config option. We can then later edit this config in real time w/ Eden dynamic config.

Raising this value may help with workflows that do a lot of IO in parallel (ex: running many ripgreps at once). This might also be used as a potential mitigation for S402137.

# This diff

This diff introduces a daemon config option for configuring the number of FUSE dispatcher threads that Eden spawns at mount creation time. This is not used yet, but it will be used in the next diff.

Reviewed By: genevievehelsel

Differential Revision: D55495411

fbshipit-source-id: 94518e971a18aee7fe90d7738c4183bb2382a60c