Disallow `className` in <Button>
Summary: Since we use stylex, providing a className will override it and be weird. By disallowing it as a prop, you're hinted to remember to migrate to `stylex` or an inline`style` instead of a classname. Maybe we'll choose to merge classNames later, but for now I don't think it's necessary.

Reviewed By: quark-zju

Differential Revision: D55509888

fbshipit-source-id: c83775956b618c3adc07d91459cd7ba6ff4b3da4