scmstore: add config to disable transaprent indexedlog syncing
Summary:
Add scmstore.sync-logs-if-changed-on-disk to configure whether we check if indexedlogs have changed on disk at read time.

By default it will be off, now, but we can turn it back on using the config.

I had to pipe the config around annoyingly. The revisionstore indexedlogutil::Store is used by various things, and I didn't want to duplicate the config check all over the place, so piped the config to a more central spot. I want to refactor some of the indexed log config stuff to get rid of IndexedLogHgIdDataStoreConfig and instead use the config more directly, but not doing that right now.

Reviewed By: quark-zju

Differential Revision: D55534005

fbshipit-source-id: f3f07cde1cd754ed66a6f7f62c8fa59db68fa870