update _strip_stack_information() to recognize \r\n (#863)
Summary:
[github] update _strip_stack_information() to recognize \r\n

By default, Sapling programmatically writes the PR body on GitHub
and will continue to overwrite it every time the user runs
`sl pr submit`. In so doing, Sapling always uses `\n` as the
line terminator for the content it generates.

However, if the user has `github.preserve-pull-request-description=true`
set and edits the pull request body on GitHub, then the next time
Sapling pulls down the pull request body to parse it, it will find that
the original line terminators it wrote will have been changed to `\r\n`.

Previous to this commit, `_strip_stack_information()` always looked
for `_HORIZONTAL_RULE` and `_SAPLING_FOOTER_MARKER` joined with `\n`,
which failed to match the newly written version containing `\r\n`.
This meant that users with `github.preserve-pull-request-description=true`
set would end up getting an extra copy of the stack information
written to the PR body after running `sl pr submit` because the first
instance was considered the source of the PR body that Sapling was
not supposed to modify based on `github.preserve-pull-request-description=true`.

This commit changes `_strip_stack_information()` to match the pattern
using a regex that recognizes both `\n` and `\r\n` as the line terminator.

Added a doctest to `_strip_stack_information()` and verified the following
passes locally:

```
./tests/run-tests.py ./tests/test-doctest.py
```

Pull Request resolved: https://github.com/facebook/sapling/pull/863

Reviewed By: muirdm

Differential Revision: D55514601

Pulled By: quark-zju

fbshipit-source-id: 58ef0f686ee391deb5767c9231064b47b489af3b