Eden CLI: Introduce RedirectionState::RealDirWithData
Summary:
For redirect fixup we need to deal with the case of finding a directory instead of a symlink. Until now, if the directory is empty we just remove it and replace it, but when it's not empty (which is the typical case) we fail.

We want redirect fixup to properly handle this case (by preserving those files and then putting them in the correct symlink), so the first step is detecting it which is what this diff adds.

Reviewed By: christhoperbernardmeta

Differential Revision: D55444550

fbshipit-source-id: 29f086b1809d0be843840463cedc99091e47ac11