Use temporary file to feed stdin to hg in tests
Summary:
Previously, eden tests use Python's subprocess module's `input=` feature to
provide stdin to hg. However, we have seen test hanging when `hg` tries to
read from stdin on macOS CI when multiple tests are run together.

Under the hood, the `input=` is using `stdin=PIPE` and threads to feed the
stdin, which seems to me that if the main Python interp (for the test runner,
not hg) is busy elsewhere then it might not be able to feed stdin.

This diff switchs the stdin handling from the PIPE+thread to a temporary file,
so it does not rely on the Python interp to feed the content and hopefully
help stablizing the tests.

Reviewed By: jdelliot

Differential Revision: D55602456

fbshipit-source-id: 84d511657f6868d74d82e27f8af6fa146edd22d1