Show inline GitHub comments and suggested changes
Summary:
In addition to top level comments on the PR, let's also render inline comments. This makes the comment button more useful. Especially since you can see suggestions and click on filenames to go there in vscode.

It's a little funky becuase the API just puts the suggested change in the HTML as a <table>, and doens't seem to expose it elsewhere. This means we need to hardcode some styles to restyle the suggested change to look decent.

One small known issue: the "count" of comments on the comment button is for top-level PR comments, not inline "review" comments. We may want to change our definition of number of comments to match.

Reviewed By: quark-zju

Differential Revision: D55620134

fbshipit-source-id: 17f456b2948498912754dcf501d7da81d6b48662