Integration test highlighting tag fetch failure
Summary: As a performance optimization, the current Mononoke Git server returns all the tags present at the server regardless of what is requested by the client. In most cases this is fine, but if the clone is shallow or restricted to certain branches that do not include the tagged commit then the response payload could end up sending a tag that points to a commit which is not part of the final packfile. This leads to failed clone with the git client erroring out. This integration test helps highlight this scenario which will be fixed in a follow up diff.

Reviewed By: mitrandir77

Differential Revision: D55559659

fbshipit-source-id: f0bbdf4f7ca2a0443fd776c2c4e581b95ebbe235