Fix tag stream fetch during git clone / pull
Summary: As detailed in the previous diff, when fetching the tag stream for git clone or pull, we end up sending more data than requested because its less work / faster to do so. However, in certain cases this could lead to invalid clone with the git client erroring out with a fatal failure (reproduced in integration test in previous diff). This diff fixes that problem by ensuring that we only fetch the set of tags are required by the git client as part of the request parameters.

Reviewed By: mitrandir77

Differential Revision: D55561758

fbshipit-source-id: 27c733f6a6d1bcf0e2e1764c90ff21a8d108435b