Implement support for ref-in-want feature in git server
Summary: The `ref-in-want` feature in Git allows to specify refs with name instead of the object id that they point to for fetching the response. This is used in cases when there is a possibility of a lag w.r.t. the bookmarks between the different servers. We are providing support for it here since the protocol includes it, however with Mononoke as a Git server there will be no leader-follower setup, i.e. all servers will have the same view of the bookmarks so this feature won't really be relevant.

Reviewed By: mitrandir77

Differential Revision: D55599625

fbshipit-source-id: c63cec1e0794be35092145ab997e0b613247f48f