Update test for implicit deletes in submodule repo
Summary:
## This stack
See D55064943.

## This diff
After going over it again, this test case is already working as expected.
It asserts that any implicit deletes in the submodule repo (repo B) is properly expanded. But the implicit deletes in the original commit will also generate implicit deletes in the expansion.

I updated the comments and also added another assertion that the files that are supposed to be deleted are not in the working copy.

Reviewed By: mitrandir77

Differential Revision: D55364555

fbshipit-source-id: fe3b0c697b4a368805905d88d61a71e0d1b81d35