Move RequestContext and RequestContextMiddleware from edenapi_service to gotham_ext
Summary: The `RequestContext` state and the `RequestContextMiddleware` were placed in `edenapi_service/middleware` crate instead of `gotham_ext` crate since `RequestContext` was deemed to be `EdenAPI` specific. However, on deeper inspection we can see that the logic contained in that file is very generic can be applicable to other HTTP services in Mononoke (e.g. Git server). Instead of reinventing the same thing for the Mononoke Git Service, I am moving this file to the `gotham_ext` crate so it can be used across projects

Differential Revision: D55692323

fbshipit-source-id: afd2a900fc49bf3f506ffe8656b30cf2ef217395