bookmark movement: propagate bookmark update log id
Summary:
This is separate from the parent diff to highlight its significance:
We want to know the log id as the outcome of bookmark movements.

In the next diff, we will use it to do away with backsyncing in Mononoke and
minimize the thundering herd effect when the backsync is lagging.

Reviewed By: gustavoavena

Differential Revision: D55133278

fbshipit-source-id: bb821fc5e5143505f4b4ac882143881c67b0273e