Replace string types with rust enums for response headers
Summary: In git protocol, there are two git headers that we need to send as part of the response. The `ServiceType`, which indicates if we are responding to a `upload-pack` request or a `receive-pack` and `ResponseType` which indicates if the response is data or set of capabilities. Till now, these were stringified since I did not know how many possible values exist for them. Now that I know, I have converted them into an enum to make it more rusty

Differential Revision: D55694178

fbshipit-source-id: 8347a1dc180d154363edbee254b4c83aa720ab36