workingcopy: do not take .git/index.lock when calculating git status
Summary:
We got reports like:

  fatal: Unable to create '.../.git/index.lock': File exists.

It seems to be caused by `git status`. Ideally git can just wait for the lock
to release itself like sl, instead of erroring out directly. But for now it
seems we need a workaround.

According to `man git-status`:

    BACKGROUND REFRESH
       By default, git status will automatically refresh the index, updating
       the cached stat information from the working tree and writing out the
       result. Writing out the updated index is an optimization that isn’t
       strictly necessary (status computes the values for itself, but writing
       them out is just to save subsequent programs from repeating our
       computation). When status is run in the background, the lock held during
       the write may conflict with other simultaneous processes, causing them
       to fail. Scripts running status in the background should consider using
       git --no-optional-locks status (see git(1) for details).

I verified that:
- `strace -f git status 2>&1 | grep index.lock` shows the index.lock is created
  then deleted.
- `strace -f git --no-optional-locks status 2>&1 | grep index.lock` shows
  nothing.

This diff changes the `git status` to pass the `--no-optional-locks` to avoid
interfering with other git commands, at the cost of losing optimization for
performance of the next `status` call. Practically, Git repos running in this
dotgit mode are hopefully small-ish so the trade-off seems okay.

Reviewed By: evangrayk

Differential Revision: D55723219

fbshipit-source-id: 4923d1da3fcfee6a5c81c49379160a22f7f696d5