Make Commit avatar draggable by increasing DraggableCommit width
Summary:
Simple semi-hacky fix to make the Avatar draggable. We shift the <DraggableCommit> left 30 pixels, but give it 30px left padding. So it renders in the same place, but 30px left of it is now considered a draggable part of it.

The alternative that may be more "correct" would be to make the <Avatar> also get wrapped in <DraggableCommit>, either by rendering Avatar + Commit together instead of handlign that in Render Dag, or by having two <DraggableCommit> containers.

I think this hacky approach is not all that bad. The drag preview is done by the brower just dragging the rendered pixels, and that magically includes the avatar here.

Reviewed By: quark-zju

Differential Revision: D55761149

fbshipit-source-id: 28dc404de052fa61f282f330e5f8cc09889061b7