Disallow selection "You are here"
Summary:
Some changes to selection made it possible to click on "you are here" and select it as a commit (since it's now a virtual commit in the dag). Let's not allow this, since it's not really what you might expect to happen visually.

Double click to open the commit in the sidebar is a nice feature, I think it's only the selection bg we don't want.

Reviewed By: quark-zju

Differential Revision: D55761965

fbshipit-source-id: 2aca8fc570f712eef295cee31193d7a963161b27