repo_metadata_logger: add a mononoke app to replace the python script
Summary:
Context: repo_metadata_logger (https://www.internalfb.com/code/fbsource/[47c9910c86bc]/fbcode/scm/repo_metadata_logger/main.py) is a python script that requests metadata about all files and directories in repos from SCS and then logs it. As this script powers the dim_scm_repo_metadata hive table which is used for tracking developer productivity, we would like to make it faster and more reliable, and a good step towards this is to turn it into a Mononoke app.

`repo_metadata` is implemented as one of the repo features, it's a method that returns a stream of `MetadataItem`s for each repo. The `repo_metadata_logger` binary goes through the metadata returned from this stream and outputs it to stdout.

For now `repo_metadata` returns a single `MetadataItem::Unknown` for each bookmark configured for the repo. Each subsequent diff will add one variant of `MetadataItem` until all types are covered and then `MetadataItem::Unknown` will be deleted.

Differential Revision: D55690658

fbshipit-source-id: 3043e29b6974e38a90610f6c5f0ad0598495e79e