repo_metadata_logger: calculate metadata for directories
Summary:
Context: repo_metadata_logger (https://www.internalfb.com/code/fbsource/[47c9910c86bc]/fbcode/scm/repo_metadata_logger/main.py) is a python script that requests metadata about all files and directories in repos from SCS and then logs it. As this script powers the dim_scm_repo_metadata hive table which is used for tracking developer productivity, we would like to make it faster and more reliable, and a good step towards this is to turn it into a Mononoke app.

Implements the logic for calculating metadata for directories. We start by getting the fsnode and unode for all files and directories in the repo (using the logic added in D55690659) and processing each. All the metadata that we need to log is either stored in fsnodes or unodes, or can be fetched using the content id from fsnodes or the unode id.

Reviewed By: RajivTS

Differential Revision: D55692093

fbshipit-source-id: 31eba002bb8c67de090da5cbf85c70681d9b4291