repo_metadata_logger: add repo config to specify bookmarks to log metadata for
Summary:
Context: repo_metadata_logger (https://www.internalfb.com/code/fbsource/[47c9910c86bc]/fbcode/scm/repo_metadata_logger/main.py) is a python script that requests metadata about all files and directories in repos from SCS and then logs it. As this script powers the dim_scm_repo_metadata hive table which is used for tracking developer productivity, we would like to make it faster and more reliable, and a good step towards this is to turn it into a Mononoke app.

Currently which bookmarks to log metadata for is hardcoded for each repo in the python script. Let's make it a repo config instead

Reviewed By: RajivTS

Differential Revision: D55426907

fbshipit-source-id: 3b7f53a3f1f6ce7634377bfb3af82eb538e8b6cf