repo_metadata_logger: create fbpkg for repo_metadata_logger
Summary:
Context: repo_metadata_logger (https://www.internalfb.com/code/fbsource/[47c9910c86bc]/fbcode/scm/repo_metadata_logger/main.py) is a python script that requests metadata about all files and directories in repos from SCS and then logs it. As this script powers the dim_scm_repo_metadata hive table which is used for tracking developer productivity, we would like to make it faster and more reliable, and a good step towards this is to turn it into a Mononoke app.

Creates fbpkg for repo_metadata_logger to be able to either run it in chronos or deploy it to tupperware.

Differential Revision: D55750064

fbshipit-source-id: f03e238605aa9077188de1bb402f875627f57f2c