Fix github pr badge color
Summary:
I noticed when publishing version 0.1.45 that github PR badges were colored incorrectly. This is due to the change to make PR badges render using <Link> instead of <VSCodeLink>. With stylex, we see an unfortunate style priority order issue. In prod builds, the vscode webview toolkit css file resets `a, a code { color: var(--vscode-link-color) }`. This has more priority than the stylex style setting the color to white in the diff badge.

To fix this, let's use a CSS style rule for this instead.

Reviewed By: quark-zju

Differential Revision: D55760185

fbshipit-source-id: b89bdfd744191babce81be0272e48e1997c0bbb1