Prevent actually selecting virutal commit
Summary:
D55761965 prevented visually selecting the "you are here" virtual commit. But we actually need to prevent it from being added to the selected state in the first place. Otherwise, things that use the selection will try to use the virtual commit, which is usually wrong.

For example, if you click "you are here", then cmd-click another commit, ISL would crash because the virtual commit was not in the dag.

Reviewed By: quark-zju

Differential Revision: D55762893

fbshipit-source-id: 05277634f5ef965004196861dc81cc7e92cf330f