tests: add a way to start EdenFS in .t tests
Summary:
This allow EdenFS to be started on .t tests, actual functionality will be added in other diffs.

At the moment this only allows a couple of tests to be run in this configuration, one for making sure that the tests can be run by CI (`test-basic.t`) and another one to verify that EdenFS can actually start(`test-eden-clone.t`) that immediately fails.

Reviewed By: muirdm

Differential Revision: D54032852

fbshipit-source-id: 6141282c7cf4d759a67915f44982a9cae38071ab