tests: add additional support for running .bat files on Windows
Summary:
On D55272594 support for running EdenFS commands through the CLI. However, there is a small issue on Windows since on that diff a script for running EdenFS is generated. On POSIX adding a suffix for scripts is not necessary, but on Windows we need to create `.bat` files.

Before this diff, debugruntest was expecting all kind of executables on Windows to be `.exe`s, but `.bat` were not supported.

Reviewed By: muirdm

Differential Revision: D55272599

fbshipit-source-id: e9926c4347570545bdeb6577ea5df63278b064f4