Ignore unknown ShardedService variants
Summary: As a SEV follow up for S405321, we want to make the parsing of `ShardedService` more resistant to unknown enum values. The rationale behind this change is that if the parsing code doesn't understand a particular variant then the rest of the application cannot make use of that variant so it should be safe to ignore it.

Reviewed By: YousefSalama

Differential Revision: D55795442

fbshipit-source-id: f74bba3fa9967073bbf0bbc2cd715eae0b405a09