commit_graph: implement segmented_slice_ancestors
Summary:
A common approach used in Mononoke when wanting to partition changesets into roughly equal parts is to slice using generation numbers so that each slice contains changesets that have generation numbers inside of a range. This is implemented in `slice_ancestors` and works well for most repos except those that have repo merges, for which the lower generation numbers end up having a higher concentration of changesets. This is relevant when using parallelized derived data backfilling because its total execution time is relative to the size of the largest slice.

This diff implements an alternative slicing strategy that first calls `ancestors_difference_segments` to convert the portion of the commit graph we want to slice into segments, then distributes the segments to slices so that all slices except potentially the last one have exactly the same size.

This additionally returns `BoundaryChangesets` which are any changesets in a slice that are parents of changesets in another slice. To try to minimize the number of these boundary changesets we first sort the segments returned from `ancestors_difference_segments` in dfs order.

Reviewed By: markbt

Differential Revision: D53518841

fbshipit-source-id: 6211f4b7eb293082526b804229dde35b8f6d7fd9