tests: create a way for running eden commands from the CLI
Summary:
There are a number of things that the EdenFS integration tests do, like starting/restarting EdenFS and running EdenFS commands. Running commands other than start/stop is not much of an issue if the proper environment variables plus CLI args are supplied for running EdenFS commands. What's more, there is a way of getting this through the EdenFS integration tests.

This diff adds a wrapper for running EdenFS on this way and makes it available through `#require`s on debugruntest tests.

Reviewed By: muirdm

Differential Revision: D55272594

fbshipit-source-id: a697663d725df8e4e0d998dba406f99ab46b8dd4