Allow cancelling highlighting tasks
Summary:
Shifting syntax highlighting to web workers broke cancelling via cancellation tokens.

We now need to send a message to the web worker to cancel, and have the worker know to use cancel messages to cancel known cancellation tokens.

One subtle thing here is that this now requires an explicit queue of work items. Before, we implicitly queued via the javascript event queue. Because we used async work, new messages would be handled as new promise tasks. But if we used this system, then the cancellation message would also get queued up, and wouldn't be processed until after the original task is done, making the cancel useless.

Instead, we make an explicit work queue, and have cancelling NOT get queued up in this way. When we finally dequeue a task to run, its cancellation token will have been updated to be cancelled.

Weirdly, this requries waiting until the next tick before starting work, or else the same problem occurs (because we don't give the task queue time to empty).

Reviewed By: quark-zju

Differential Revision: D55716393

fbshipit-source-id: ca8c76bbea28e27215390f631df60173734f0c58