Load web workers in vscode
Summary:
VS Code Webviews allow Web Workers, but have a specific requirement that they are loaded via a `blob:` uri and cannot make further imports.

This means we need to change how we load the web worker code and get the data via `fetch` instead of a regular import. We can change our vite bundling step to ensure we know the worker's file name so we can fetch it consistently.

Also, by using a Blob URI, it means the "base" uri inside the worker for futher fetch calls is incorrect. This affects how it fetches the onigurama WASM file and also the textmate grammar files. This means we need to explicitly tell the worker what baseURI it should use, so it can use that in fetch calls. We can do this with a message we pass on worker setup.

Finally, in dev mode, vite will not have built the worker file in a way that works. That's because vite's dev mode assumes it can use imports and dynamically fetch other files (this is how it lazily compiles so quickly). But workers in vscode webviews may not use imports (even in "module" mode). So there's not much we can do without complicating the build system a lot. Let's just not use web workers in vscode dev mode, and use the synchronous version of the highlighting (where it does it on the main thread). This has worse perf but it's ok in dev mode I think.

Reviewed By: quark-zju

Differential Revision: D55777476

fbshipit-source-id: 8408e52264834cb09bb46078b44a720d0ae4efae