Use custom emitted manifest from vite
Summary:
The ISL server uses vite's manifest to server only specific files (in production). However, this manifest was not including outputted web worker js files. This meant we 404'd the web worker js file in production.

A way around this is to write our own manifest via a simple vite plugin. This actually simplifies the server, where we can just take this manifest literally instead of extracting filenames. And now it includes the web worker file as well.

Reviewed By: quark-zju

Differential Revision: D55720599

fbshipit-source-id: 1600610dbb8e3c15fb8318b429ac37c998c0c539