Rewrite comparison context line fetching
Summary:
Before, we fetched comparison view context lines ("expand 5 lines") via an atom family, keyed by an object.

This doesn't actually really work in jotai weakAtomFamily. The context appears and disappears and does weird stuff when the component rerenders. This is because it was using the parameter object reference to cache.

This part is one of the older parts of the codebase, and this pattern of using an atom to fetch using a param object is questionable in hindsight. Let's instead just use a simple hook which has a simple react useState and do a normal useEffect fetch.

In dev mode, sure, this calls `sl cat` twice. I think that's fine.

Reviewed By: quark-zju

Differential Revision: D55716392

fbshipit-source-id: 5d4b7c8a167c2c06f2f94eafeb32bc461943cd74