Split submodule expansion code into multiple modules
Summary:
## This stack
See D55696045

## This diff
The `git_submodules` file is becoming very large with the validation code that I'm adding.

Since I'll also be writing more code to handle backsyncing in the future, I decided to split the file into a module now, while it's still easy to do so.

This also let's me have a better defined interface to outside the `git_submodules` module.

Reviewed By: mitrandir77

Differential Revision: D55788403

fbshipit-source-id: 5d0f53b0a8300f2c82612752bc6188267baaa573