Encapsulate data needed for submodule expansion in struct
Summary:
## This stack
This stack will introduce proper validation of bonsais to make sure that the submodule expansion is always consistent with the submodule metadata file.

I'll start with some refactoring diffs to make everything easier to review.

The goal of validation is to make sure that **the submodule metadata file is always kept in sync with the expansion**.
How this will be done is more complicated than it seems, but on a high level we will
1 Get the fsnode in the submodule repo from the commit in the metadata file
2. Derive the fsnode for the new commit
3. Get the entry for the expansion and compare with the fsnode from step #1.
4. We'll perform many validations of consistency in this process (e.g. if expansion is changed, metadata file has to be changed as well).

## This diff
Encapsulates everything needed by the submodule expansion module into a struct.
I'm doing this because I don't want to keep adding more and more parameters to multiple function parameters every time I need something else.

One thing I want to make sure is that this type is cheap to clone, because it only contains references.

Reviewed By: mitrandir77

Differential Revision: D55696045

fbshipit-source-id: 85ab8756ca80a19ad561e585f301adad501fb718