Rename source_repo to small_repo
Summary:
## This stack
See D55696045

## This diff

The initial plan was to only provide submodule expansion in forward syncing, i.e. small to large. So pretty much the entire code was written under that assumption and used `source_repo` to refer to `small_repo` and `target_repo` to refer to the large repo.

Because we'll support backsyncing as well, I'm already trying to implement everything new with that in mind. Keeping references to source and target repo is very confusing and not needed, because expansion will only be from small to large.

Validation, which I'll introduce in later diffs, will be used in both directions, but the repos will need to be explicitly labeled as large and small, because the validation logic will always be the same, regardless of direction.

I also added a TODO comment to crash if we try to backsync changes with a small repo that has git submodule action set to expand.
I'll do that in next few diffs.

Differential Revision: D55738695

fbshipit-source-id: bf748a465b34cf301a887e46a82e6433df2839cb