Handle implicitly deleting a submodule
Summary:
## This stack
See D55696045

## This diff
In this case, the only work I need to do is generate a deletion for the submodule metadata file, because the expansion is already deleted implicitly.

I updated the test to confirm that we're generating the explicit deletion of the metadata file **and the working copy doesn't contain the expansion after the file is added**, because it was deleted implicitly.

Reviewed By: mitrandir77

Differential Revision: D55799048

fbshipit-source-id: 9a0dd06fea130429143445aec33050db5404e83c