Refactor unit tests to support updating commit sync config
Summary:
## This stack
See D55696045

## This diff
Refactors a few things in the unit tests to make a change to the `test_implicitly_deleting_file_with_submodule` in next diff (will be explained there).

The change is basically: Instead of always building the config using `repo_b` as a submodule dependency, require this repo to B passed as submodule dependency every time.
While this introduce the duplication of one-liners, it makes it more explicit in every test case which repos are being set as submodule dependency for each config version.

I also split the function that creates the commit syncer into 2 smaller ones that might be reused later.

Reviewed By: mitrandir77

Differential Revision: D55793698

fbshipit-source-id: 90504d2a8c2a150302c09bf5e93e7f64f3489043