Introduce submodule validation when action is expand
Summary:
## This stack
See D55696045

## This diff
This diff introduces the validation flow for submodule expansion.

For context, the validation function needs to be built **with both forward sync and backsync in mind**.
Right now, it will be used to validate that the submodule expansion is working as properly, which will give us confidence that after we use expansion in production for Phase 1, the commits will be valid and "back-syncable" when we implement backsyncing support.

One requirement I wanted to have is that **you can't use submodule expansion without validating the final bonsai**.
I don't want to depend on callers doing validation, so I encapsulated both expansion and validation into a public function and reduced visibility of the expansion to private.

In the future, I'll make the `validate_all_submodule_expansions` function public, so we can use it to validate that commits made directly in the large repo are valid.

## Next steps
Validation is already WIP and is quite complex (see V10 of D55248739 which I'm using to keep my WIP changes).
I'll try to split it as much as possible, by introducing each of the 3 steps I mentioned in the documentation in separate diffs.

These validations will be done sequentially, from the simplest to the most complex. Implementing the first check is simple and we can already fail validation if that doesn't pass.

Next diff will introduce the 1st validation.

Reviewed By: mitrandir77

Differential Revision: D55696475

fbshipit-source-id: 25479b5d505adc50422fbe04b245aa727e725d8b