evalframe: relax input type on sapling_cext_evalframe_resolve_frame
Summary:
When built with debuginfo, lldb will type check the function, and refuse to run
if the input type is not `PyFrameObject *`. After `strip -S`, lldb does not
perform the type check, but cannot recognize `PyFrameObject`.

To make it work for both debuginfo and `strip -S` case, just take `size_t` as
input.

Reviewed By: zzl0

Differential Revision: D55784492

fbshipit-source-id: 3aaafb9d478df20db133cb943c0ad94171fbf4bb