dbgutil: resume process on Windows
Summary:
I noticed that on Windows the debugee gets stuck in "Suspended"
state even after exiting lldb. This diff attempts to solve the
problem by:
- Using `process.Detach()` explicitly (which turns out ineffective)
- Using `process.Continue()` to force (?) continuing running the
  process. This turns out effective.

It seems the debugger injects a thread that triggers a breakpoint
(0x80000003 is "breakpoint") that might be the cause (?)

  thread 15: tid = 0x8030, 0x00007ff909e73061 ntdll.dll`DbgBreakPoint + 1, stop reason = Exception 0x80000003 encountered at address 0x7ff909e73060
    frame 0: 0x00007ff909e73061 ntdll.dll`DbgBreakPoint + 1
    frame 1: 0x00007ff909ea69ae ntdll.dll`DbgUiRemoteBreakin + 78
    frame 2: 0x00007ff9087e257d kernel32.dll`BaseThreadInitThunk + 29
    frame 3: 0x00007ff909e2aa58 ntdll.dll`RtlUserThreadStart + 40

Anyway, with this change it seems the debugee can resume fine.

Reviewed By: zzl0

Differential Revision: D55907874

fbshipit-source-id: 55cb5f88115b0416703740f24ccc4607afa334a2