Add context menu action to delete a local bookmark
Summary:
Allow right clicking a bookmark and deleting it.

This only applies to local bookmarks (not remote bookmarks or stable locations)

we show this menu on the bookmark's context menu, not the commits. We do this because you can have multiple bookmarks on a commit, and showing the mall in the same context menu is messy.

Note: no optimistic dag mutation in this diff, so the bookmark doesn't optimistically hide itself. I'll add that in a follow up diff.

Reviewed By: quark-zju

Differential Revision: D55938826

fbshipit-source-id: 64904045dc8b64f08763e1c412f9cbf678ac5d40